dialog-0j03_0_37
0j/03, Lib 0.37 (Manual revision 22):

Library: Choice mode.

Library: Added hooks that are queried early and late on every
tick.

Library: Undo now operates at the level of commandlines, rather
than complex actions. "UNDO" must now be typed on a line of its
own, but behaves in a less surprising way.

Library: Before entering a closed container, try opening it.

Compiler: Fixed rare bug related to variables bound inside
if-statements.

Compiler: Minor optimization of collect-into.

Debugger: Removed spurious blank line when entering a status
bar environment.

+++ src/ast.h
@@ -385,7 +385,7 @@ struct program {
 	program_ticker_t	eval_ticker;
 	int			nwordmappred;
 	struct word		**clausevars;
-	int			*clausevarcounts;
+	uint8_t			*clausevarcounts;
 	int			nclausevar;
 	int			nalloc_var;
 	char			*meta_author;
